+++ relay-filter/src/lib.rs
@@ -21,9 +21,9 @@ pub mod browser_extensions;
 pub mod client_ips;
 pub mod csp;
 pub mod error_messages;
-pub mod health_check;
 pub mod legacy_browsers;
 pub mod localhost;
+pub mod transaction_name;
 pub mod web_crawlers;
 
 mod common;
@@ -57,7 +57,7 @@ pub fn should_filter(
     browser_extensions::should_filter(event, &config.browser_extensions)?;
     legacy_browsers::should_filter(event, &config.legacy_browsers)?;
     web_crawlers::should_filter(event, &config.web_crawlers)?;
-    health_check::should_filter(event, &config.ignore_transactions)?;
+    transaction_name::should_filter(event, &config.ignore_transactions)?;
 
     Ok(())
 }
